Extract nested class to own file
