+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/statements/expressions/CollectionComprehension.kt
@@ -45,98 +45,6 @@ import org.neo4j.ogm.annotation.Relationship
  */
 class CollectionComprehension : Expression(), ArgumentHolder {
 
-    /** This class holds the variable, iterable and predicate of the [CollectionComprehension]. */
-    class ComprehensionExpression : Expression(), ArgumentHolder {
-        @Relationship("VARIABLE")
-        var variableEdge =
-            astOptionalEdgeOf<Statement>(
-                onChanged = { _, new ->
-                    val end = new?.end
-                    if (end is Reference) {
-                        end.access = AccessValues.WRITE
-                    }
-                }
-            )
-
-        /**
-         * This field contains the iteration variable of the comprehension. It can be either a new
-         * variable declaration or a reference (probably to a new variable).
-         */
-        var variable by unwrapping(ComprehensionExpression::variableEdge)
-
-        @Relationship("ITERABLE") var iterableEdge = astOptionalEdgeOf<Statement>()
-
-        /** This field contains the iteration subject of the loop. */
-        var iterable by unwrapping(ComprehensionExpression::iterableEdge)
-
-        @Relationship("PREDICATE") var predicateEdges = astEdgesOf<Statement>()
-
-        /**
-         * This field contains the predicates which have to hold to evaluate `statement(variable)`
-         * and include it in the result.
-         */
-        var predicates by unwrapping(ComprehensionExpression::predicateEdges)
-
-        override fun toString() =
-            ToStringBuilder(this, TO_STRING_STYLE)
-                .appendSuper(super.toString())
-                .append("variable", variable)
-                .append("iterable", iterable)
-                .append("predicate", predicates)
-                .toString()
-
-        override fun equals(other: Any?): Boolean {
-            if (this === other) return true
-            if (other !is ComprehensionExpression) return false
-            return super.equals(other) &&
-                variable == other.variable &&
-                iterable == other.iterable &&
-                predicates == other.predicates
-        }
-
-        override fun hashCode() = Objects.hash(super.hashCode(), variable, iterable, predicates)
-
-        override fun addArgument(expression: Expression) {
-            if (this.variable == null) {
-                this.variable = expression
-            } else if (this.iterable == null) {
-                this.iterable = expression
-            } else {
-                this.predicates += expression
-            }
-        }
-
-        override fun replaceArgument(old: Expression, new: Expression): Boolean {
-            if (this.variable == old) {
-                this.variable = new
-                return true
-            }
-
-            if (this.iterable == old) {
-                this.iterable = new
-                return true
-            }
-
-            var changedSomething = false
-            val newPredicates =
-                this.predicates.map {
-                    if (it == old) {
-                        changedSomething = true
-                        new
-                    } else it
-                }
-            this.predicates.clear()
-            this.predicates.addAll(newPredicates)
-            return changedSomething
-        }
-
-        override fun hasArgument(expression: Expression): Boolean {
-            return this.variable == expression ||
-                this.iterable == expression ||
-                expression in this.predicates
-        }
-    }
-
     @Relationship("COMPREHENSION_EXPRESSIONS")
     var comprehensionExpressionEdges = astEdgesOf<ComprehensionExpression>()
     /**
